Prototype version of rewritten azure backend
Older azure clients from sundry sources have been replaced with the official
azure-sdk-for-go library. This is a rough draft that only covers the
most basic authentication situation, and error handling is only very
roughly set. Additionally, no testing has yet been done on this.
However, this is a good place to start for checking that we're on the
right track, and now there's something to show and share.

Signed-off-by: Larry Bordowitz <laurence.bordowitz@gmail.com>