Transition to Shared Keys Authentication
In the azurerm backend, either AzureAD (EntraID) is used as the primary
authentication method, or the authentication is used as a bootstrap. By
that, I mean the authentication method is used to obtain the Shared Keys
for the Storage Account. By using the Storage Account Shared Keys, other
authorization and role binding is not needed, whereas if EntraID-based
authentication is used, you do need an additional role to access storage
blobs (for example Storage Blob Data Contributor).

We're temporarily using this method in order to maintain parity and have
a bit of an easier time while we work out the non-authentication kinks
with the rewritten backend.

Signed-off-by: Larry Bordowitz <laurence.bordowitz@gmail.com>